Support for .NET hybrid cache
Motivation
----------
ASP.NET Core 9 has added the concept of a hybrid cache. This is a mixed
memory and distributed cache, which uses memory as the L1 cache and
efficiently uses the distributed cache as the L2 cache. While the
specific implementation is in preview, the abstractions supporting it
are available now. By adding `IBufferDistributedCache` interface and an
implementation of `IHybridCacheSerializerFactory` we'll be compatible.

Modifications
-------------
- Upgrade to Couchbase SDK 3.6.5 to get support for deserializing from
  `ReadOnlySequence<T>` and serializing to `IBufferWriter<byte>` via
  the new `IBufferedTypeSerializer` interface.
- Upgrade to `Microsoft.Extensions.Caching.Abstractions` 9.0.1. Note
  that this is backward compatible to .NET 6.
- Implement an `IHybridCacheSerializer<T>` wrapper for
  `IBufferedTypeSerializer` and expose it via a new public type
  implementing `IHybridCacheSerializerFactory`.
- Add `IBufferDistributedCache` to `CouchbaseCache`.
- Add support for the new serialization and deserialization types to
  `CacheTranscoder`.
- Add integration and unit tests.

Results
-------
Consumers who want to use the preview support may do so, and we'll be
immediately ready when the final `Microsoft.Extensions.Caching.Hybrid`
package is released.