Always use System.Text.Json for serializing mutexes
Motivation
----------
In some System.Text.Json configurations of the Couchbase SDK, mutex
documents are not serialized correctly. Additionally, if using
a serializer built from a JsonSerializerContext they will fail because
the lock document isn't included in the context. Since the class for the
document is internal, it isn't possible for the consumer to include it
in their JsonSerializerContext.

Modifications
-------------
Always use a JsonTranscoder that uses an internal System.Text.Json
serializer built around an internal JsonSerializerContext for all mutex
documents.

Results
-------
- Consistent behavior regardless of the serializer selected in the
  Couchbase SDK or its configuration
- This should also offer higher performance than the Newtonsoft
  serializer or reflection-based System.Text.Json serializer

Fixes #118