Correctly encode ReadOnlySequence<byte> when caching
Motivation
----------
When ReadOnlySequence<byte> is encountered by the CacheTranscoder it is
being written to the document twice. First by the CacheTranscoder, then
again by the default transcoder configured on the Couchbase SDK.

This is particularly problematic when trying to use the new .NET
HybridCache, which writes using ReadOnlySequence<byte>, and effectively
makes hybrid caching with Couchbase non-functional. It will consistently
fail to fetch from Couchbase with an InvalidData error logged.

Modifications
-------------
- Properly short-circuit when encoding the type in the CacheTranscoder
- Update to use a released version of HybridCaching. This is why the
  problem was missed originally, it worked on pre-release versions.
- Add hybrid caching to the example application.